fix(runtime): project completion bridge status
